consistent tokens docs
