+++ docs/tokens/borders/tailwind-examples/border-radius.stories.tsx
@@ -3,11 +3,11 @@ import { Meta, StoryFn } from "@storybook/react"
 import classnames from "classnames"
 import { kaizenTailwindTheme } from "@kaizen/tailwind"
 import { TailwindStoryTemplate } from "~storybook/components/TailwindStoryTemplate"
-import { utilityDescription } from "../../../../helpers/utilityDescription"
+import { utilityDescription } from "../../../helpers/utilityDescription"
 
-const prefix = "shadow-"
+const prefix = "rounded-"
 const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
-  Object.entries(kaizenTailwindTheme.boxShadow || []).map(
+  Object.entries(kaizenTailwindTheme.borderRadius || []).map(
     ([suffix, cssProperty]) => ({
       utilityClassName: `${prefix}${suffix}`,
       cssProperty,
@@ -15,25 +15,32 @@ const classEntries: Array<{ utilityClassName: string; cssProperty: string }> =
   )
 
 export default {
-  title: "Tier 1/Tokens/Colour/Tailwind/Box Shadow",
+  title: "Tier 1/Tokens/Borders/Tailwind/Border Radius",
   parameters: {
     a11y: { disable: true },
     chromatic: { disable: false },
     docsLayout: "fullPage",
     docs: {
       description: {
-        component: utilityDescription(prefix, classEntries[0].utilityClassName),
+        component: utilityDescription(prefix, classEntries[3].utilityClassName),
       },
     },
   },
 } satisfies Meta
 
-export const BoxShadow: StoryFn<{ isReversed: boolean }> = ({ isReversed }) => (
+export const BorderRadius: StoryFn<{ isReversed: boolean }> = ({
+  isReversed,
+}) => (
   <TailwindStoryTemplate
-    compiledCssPropertyName="box-shadow"
+    compiledCssPropertyName="border-radius"
     classKeyValues={classEntries}
     renderExampleComponent={(utilityClass): React.ReactElement => (
-      <div className={classnames("w-[100px] h-[100px]", utilityClass)} />
+      <div
+        className={classnames(
+          "w-[100px] h-[100px] border border-purple-500",
+          utilityClass.replace("-DEFAULT", "")
+        )}
+      />
     )}
     isReversed={isReversed}
   />
