+++ docs/tokens/shadow/shadow.mdx
@@ -1,0 +1,7 @@
+import { Meta } from "@storybook/blocks"
+
+<Meta title="Tier 1/Tokens/Shadow" />
+
+# Shadow
+
+ TBC
