+++ docs/tokens/shadow/tailwind-examples/shadow.mdx
@@ -3,7 +3,7 @@ import * as Examples from "./shadow.stories"
 
 <Meta of={Examples} />
 
-# Box Shadow
+# Shadow
 
 <Description />
