+++ docs/tokens/tokens.mdx
@@ -18,13 +18,15 @@ There are many ways to access the tokens:
 
 #### Tailwind
 
-Colours are accessible via normal tailwind prefixes ([example: text](https://tailwindcss.com/docs/text-color)) eg. `bg-*`, `text-*`, `border-*`
+For example, Colours are accessible via normal tailwind prefixes ([example: text](https://tailwindcss.com/docs/text-color)) eg. `bg-*`, `text-*`, `border-*`
 
 ```html
 <div class="bg-blue-500">
   ...
 </div>
 ```
+
+You can find more examples in each category of tokens.
 
 #### SCSS
 ```scss
